Update root route to be practica
